Ensure that parent scopes are inserted to `childBlocks` map.
Previously the block child mapping consisted only of blocks that are
directly in some CFG nodes. We need to also include blocks that are
the enclosing scope of CFG nodes of a subgraph. Without this, blocks
corresponding to scopes such as the top-level scope, will not have
the corresponding child mapping.