Fix exception handling in Async/Await implementation.
Problem was that the promise was prematurely failed in the `async`
coroutine, instead of passing control to user code.

This revealed a bug in the initial exception check. The exception
check had to be pushed deep down into the deepest `try` block of the
entrypoint -- without this, the exception from a direct call was just
being rethrown at the beginning of the method, regardless of the error
handlers in the body of the method.